requirements.txt
asgiref==3.3.1
cycler==0.10.0
Django==3.1.5
djangorestframework==3.12.2
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
